Use pip_install_upper_constraints for plugin deps
This commit allows including OSA global constraints when installing tempest
plugins.  In gating, this includes openstack-ansible's
global-requirement-pins.txt via a URL.  In a full deployment, it includes
requirements_absolute_requirements.txt as retrieved from the repo server, to
which openstack-ansible/global-requirement-pins.txt was accounted for in the
repo build process.